Move helper functions to utils.js
