+++ treasury-fees-monitor/src/agent.js
@@ -1,11 +1,9 @@
 const {
   Finding, FindingSeverity, FindingType, ethers, getEthersProvider,
 } = require('forta-agent');
-const fs = require('fs');
-const csv = require('csv-parser');
 const BigNumber = require('bignumber.js');
 
-const { getAbi } = require('./utils');
+const { getAbi, calculateStatistics, parseCsvAndCompute } = require('./utils');
 
 // load any bot configuration parameters
 const config = require('../bot-config.json');
@@ -34,60 +32,6 @@ function createAlert(protocolName, protocolAbbrev, developerAbbrev, type, severi
       tokenPriceEth: tokenPriceEth.toString(),
       premiumEth: premiumEth.toString(),
     },
-  });
-}
-
-function calculateStatistics(currMean, currVariance, currNumDataPoints, newValue) {
-  let newMean = 0;
-  let newStdDev = 0;
-  let newVariance = 0;
-  let newNumDataPoints = currNumDataPoints + 1;
-
-  if (currNumDataPoints === 0) {
-    newMean = newValue;
-    newNumDataPoints = 1;
-  } else {
-    newMean = (currMean * (currNumDataPoints / newNumDataPoints)) + (newValue / newNumDataPoints);
-    newVariance = (
-      (((currVariance * currNumDataPoints) + ((newValue - newMean) * (newValue - currMean)))
-        / newNumDataPoints)
-    );
-    newStdDev = Math.sqrt(newVariance);
-  }
-
-  return {
-    mean: newMean,
-    stdDev: newStdDev,
-    variance: newVariance,
-    numDataPoints: newNumDataPoints,
-  };
-}
-
-async function parseCsvAndCompute(csvFileName, tokenInfo, tokenPriceInfo) {
-  return new Promise((resolve, reject) => {
-    let mean = 0;
-    let stdDev = 0;
-    let variance = 0;
-    let numDataPoints = 0;
-
-    fs.createReadStream(`${__dirname}/${csvFileName}`)
-      .pipe(csv())
-      .on('data', (data) => {
-        // scale the premium for each row of data in the CSV file by the given asset's decimals
-        const denominator = (new BigNumber(10)).pow(tokenInfo[data.asset]);
-        const scaledPremium = parseFloat(
-          (new BigNumber(data.premium)).div(denominator),
-        );
-        const premiumInEth = scaledPremium * tokenPriceInfo[data.asset];
-
-        ({
-          mean, stdDev, variance, numDataPoints,
-        } = calculateStatistics(mean, variance, numDataPoints, premiumInEth));
-      })
-      .on('end', () => resolve({
-        mean, stdDev, variance, numDataPoints,
-      }))
-      .on('error', () => reject());
   });
 }
